runtime: add comments
