+++ daemon/README.md
@@ -1,2 +1,2 @@
-# bpf-mocker
+# bpf_tracer
