+++ runtime/src/bpf_helper.cpp
@@ -148,6 +148,7 @@ uint64_t bpf_get_stack(uint64_t, uint64_t buf, uint64_t sz, uint64_t, uint64_t)
 	memset((void *)(uintptr_t)buf, 0, sz);
 	return sz;
 }
+
 uint64_t bpf_ktime_get_coarse_ns(uint64_t, uint64_t, uint64_t, uint64_t,
 				 uint64_t)
 {
@@ -155,6 +156,7 @@ uint64_t bpf_ktime_get_coarse_ns(uint64_t, uint64_t, uint64_t, uint64_t,
 	clock_gettime(CLOCK_MONOTONIC_COARSE, &ts);
 	return (uint64_t)ts.tv_sec * 1000000000 + ts.tv_nsec;
 }
+
 uint64_t bpf_ringbuf_output(uint64_t rb, uint64_t data, uint64_t size,
 			    uint64_t flags, uint64_t)
 {
@@ -184,6 +186,7 @@ uint64_t bpf_ringbuf_reserve(uint64_t rb, uint64_t size, uint64_t flags,
 	}
 	return (uint64_t)(uintptr_t)bpftime_ringbuf_reserve(fd, size);
 }
+
 uint64_t bpf_ringbuf_submit(uint64_t data, uint64_t flags, uint64_t, uint64_t,
 			    uint64_t)
 {
@@ -208,6 +211,7 @@ uint64_t bpf_ringbuf_discard(uint64_t data, uint64_t flags, uint64_t, uint64_t,
 	bpftime_ringbuf_submit(fd, (void *)(uintptr_t)data, true);
 	return 0;
 }
+
 uint64_t bpf_perf_event_output(uint64_t ctx, uint64_t map, uint64_t flags,
 			       uint64_t data, uint64_t size)
 {
